quick setup migrations

+++ database/migrations/2018_12_25_021847_create_purchases_table.php
@@ -14,7 +14,7 @@ class CreatePurchasesTable extends Migration
     public function up()
     {
         Schema::create('purchases', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('invoice_number', 20);
             $table->date('date');
             $table->double('net_amount', 15, 2);
@@ -26,9 +26,9 @@ class CreatePurchasesTable extends Migration
             $table->foreign('company_branch_id')->references('id')->on('company_branches')->onDelete('cascade');
             $table->Integer('company_id')->unsigned()->nullable();
             $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
-            $table->integer('creator_user_id')->unsigned()->nullable();
+            $table->bigInteger('creator_user_id')->unsigned()->nullable();
             $table->foreign('creator_user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('updator_user_id')->unsigned()->nullable();
+            $table->bigInteger('updator_user_id')->unsigned()->nullable();
             $table->foreign('updator_user_id')->references('id')->on('users')->onDelete('cascade');
             $table->softDeletes();
             $table->timestamps();
