+++ database/migrations/2018_12_25_022606_create_purchase_items_table.php
@@ -15,7 +15,7 @@ class CreatePurchaseItemsTable extends Migration
     {
         Schema::create('purchase_items', function (Blueprint $table) {
             
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->Integer('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->double('unit_price', 15, 2);
@@ -25,7 +25,7 @@ class CreatePurchaseItemsTable extends Migration
             $table->date('manufactured_date')->nullable();
             $table->date('expiry_date')->nullable();
             $table->double('net_amount', 15, 2);
-            $table->Integer('purchase_id')->unsigned();
+            $table->bigInteger('purchase_id')->unsigned();
             $table->foreign('purchase_id')->references('id')->on('purchases')->onDelete('cascade');
             $table->softDeletes();
             $table->timestamps();
