+++ database/migrations/2018_12_25_023904_create_sales_table.php
@@ -14,7 +14,7 @@ class CreateSalesTable extends Migration
     public function up()
     {
         Schema::create('sales', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->double('net_amount', 15, 2);
             $table->double('discount', 8, 2);
             $table->double('extra_discount', 8, 2);
@@ -28,9 +28,9 @@ class CreateSalesTable extends Migration
             $table->foreign('company_branch_id')->references('id')->on('company_branches')->onDelete('cascade');
             $table->Integer('company_id')->unsigned()->nullable();
             $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
-            $table->integer('creator_user_id')->unsigned()->nullable();
+            $table->bigInteger('creator_user_id')->unsigned()->nullable();
             $table->foreign('creator_user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('updator_user_id')->unsigned()->nullable();
+            $table->bigInteger('updator_user_id')->unsigned()->nullable();
             $table->foreign('updator_user_id')->references('id')->on('users')->onDelete('cascade');
             $table->softDeletes();
             $table->timestamps();
